Scripts fixes for p2p and peer installation on windows

+++ scripts/public/dev/launcher-p2p-install-windows.py
@@ -8,20 +8,32 @@ import stat
 
 def subutaistart():
 
-    subutai.AddStatus("Downloading p2p binary")
-
     tmpDir = subutai.GetTmpDir()
     installDir = subutai.GetInstallDir()
 
-    subutai.download("nssm.exe")
-    while subutai.isDownloadComplete() != 1:
-        sleep(0.05)
-
     subutai.AddStatus("Installing Service Manager")
-    subutai.UnregisterService("Subutai P2P")
-    subutai.ProcessKill("nssm.exe")
-    copyfile(tmpDir+"nssm.exe", installDir+"bin/nssm.exe")
 
+    if os.path.exists(installDir+"nssm.exe"):
+        subutai.UnregisterService("Subutai P2P")
+        sleep(30)
+        subutai.ProcessKill("nssm.exe")
+        subutai.download("nssm.exe")
+        while subutai.isDownloadComplete() != 1:
+            sleep(0.05)
+
+        copyfile(tmpDir+"nssm.exe", installDir+"bin/nssm.exe")
+
+    if not os.path.exists(installDir+"nssm.exe"):
+        subutai.download("nssm.exe")
+        while subutai.isDownloadComplete() != 1:
+            sleep(0.05)
+
+        subutai.UnregisterService("Subutai P2P")
+        sleep(30)
+        subutai.ProcessKill("nssm.exe")
+        copyfile(tmpDir+"nssm.exe", installDir+"bin/nssm.exe")
+
+    subutai.AddStatus("Downloading p2p binary")
     subutai.download("p2p.exe")
     while subutai.isDownloadComplete() != 1:
         sleep(0.05)
@@ -36,6 +48,9 @@ def subutaistart():
     if os.path.exists(installDir+"bin/p2p.exe"):
         subutai.AddStatus("Removing previous installation")
         os.remove(installDir+"bin/p2p.exe")
+
+    if os.path.exists(installDir+"bin/p2p.exe"):
+        subutai.RaiseError("Failed to remove p2p binary")
 
     subutai.AddStatus("Copying files")
     copyfile(tmpDir+"p2p.exe", installDir+"bin/p2p.exe")
